+++ src/js/utils/isIOS.js
@@ -1,0 +1,7 @@
+let testResult = null;
+export const isIOS = () => {
+    if (testResult === null) {
+        testResult = /iPad|iPhone|iPod/.test(navigator.userAgent) && !window.MSStream;
+    }
+    return testResult;
+}